chore: move year selector to the separate component
